Fixed problem with selection
Since adding the horizonal split pane it was getting a value of 0
for "this.__screen.topOffset".  Now the box for the actual widget
is found with "event.getTarget().getContainerLocation("box")".
This is only valid after the widget is displayed on the screen.
It is set in the listener for the "appear" event and then the
listener for "resize" is set up.